requirements.txt
llama-index==0.10.11
google-generativeai==0.3.2
llama-index-llms-gemini==0.1.4
llama-index-embeddings-gemini==0.1.4
python-dotenv==1.0.1
